[REF] l10n_br_account: clean up inherited fields
Some fields are defined in a superficial way in the inherited
move form that is considered specific for mobile edition by Odoo but not
for us in the Brazilian localization. To make it work properly
we had to define some extra attributes just like in the editable tree view.

However, in v14 some of these attributes where added in the inherited form
view so they can be removed from the view extension as we do here.